runtime: fix struct Sigaction for Linux/ARM         if we were to use sizeof(sa.sa_mask) instead of 8 as the last argument         to rt_sigaction, we would have already fixed this bug, so also updated         Linux/386 and Linux/amd64 files to use that; also test the return value         of rt_sigaction.
R=dave, rsc
CC=golang-dev
https://golang.org/cl/6297087

+++ src/pkg/runtime/defs_linux_arm.h
@@ -143,6 +143,6 @@ struct Sigaction {
 	void *sa_handler;
 	uint32 sa_flags;
 	void *sa_restorer;
-	uint32 sa_mask;
+	uint64 sa_mask;
 };
 #pragma pack off
